Fix file diractories
