+++ include/mpc/control/alternative.hpp
@@ -1,8 +1,7 @@
 /// @file alternative.hpp
 #pragma once
 #include <mpc/control/applicative.hpp>
-#include <mpc/control/basic_traits.hpp>
-#include <mpc/functional/perfect_forward.hpp>
+#include <mpc/prelude.hpp>
 
 namespace mpc {
   // alternative
