+++ include/mpc/control/trans/class.hpp
@@ -1,7 +1,6 @@
 /// @file class.hpp
 #pragma once
-#include <functional> // std::invoke
-#include <mpc/functional/perfect_forward.hpp>
+#include <mpc/prelude.hpp>
 
 namespace mpc {
   // monad_trans
