+++ include/mpc/data/char.hpp
@@ -2,7 +2,7 @@
 #pragma once
 #include <cctype>
 #include <string>
-#include <mpc/functional/perfect_forward.hpp>
+#include <mpc/prelude.hpp>
 
 namespace mpc {
   namespace detail {
