+++ include/mpc/data/maybe.hpp
@@ -1,10 +1,9 @@
 /// @file maybe.hpp
 #pragma once
-#include <functional>
 #include <variant>
 #include <mpc/control/alternative.hpp>
 #include <mpc/control/monad.hpp>
-#include <mpc/functional/perfect_forward.hpp>
+#include <mpc/prelude.hpp>
 #include <mpc/utility/alternative_value_t.hpp>
 #include <mpc/utility/single.hpp>
