+++ include/mpc/prelude.hpp
@@ -1,0 +1,13 @@
+/// @file prelude.hpp
+#pragma once
+#include <functional>         // std::invoke
+#include <mpc/functional.hpp> // perfect_forward
+
+#include <mpc/prelude/compose.hpp>
+#include <mpc/prelude/constant.hpp>
+#include <mpc/prelude/flip.hpp>
+#include <mpc/prelude/fst.hpp>
+#include <mpc/prelude/id.hpp>
+#include <mpc/prelude/infix.hpp>
+#include <mpc/prelude/operations.hpp>
+#include <mpc/prelude/unit.hpp>
