+++ include/mpc/utility.hpp
@@ -2,7 +2,6 @@
 #pragma once
 #include <mpc/utility/alternative_value_t.hpp>
 #include <mpc/utility/copyable_box.hpp>
-#include <mpc/utility/unit.hpp>
 #include <mpc/utility/overloaded.hpp>
 #include <mpc/utility/single.hpp>
 #include <mpc/utility/tuple_like.hpp>
