+++ tests/CMakeLists.txt
@@ -71,7 +71,6 @@ FetchContent_MakeAvailable(Catch2)
 add_subdirectory(control)
 add_subdirectory(control/trans)
 add_subdirectory(data)
-add_subdirectory(draft)
-add_subdirectory(functional)
+add_subdirectory(prelude)
 add_subdirectory(type_traits)
 add_subdirectory(utility)
