+++ tests/data/data.cpp
@@ -2,8 +2,6 @@
 #include <catch2/catch.hpp>
 #include <mpc/control.hpp>
 #include <mpc/data.hpp>
-#include <mpc/functional/infix.hpp>
-#include <mpc/functional/operations.hpp>
 #include "../stdfundamental.hpp"
 
 TEST_CASE("data Identity", "[data]") {
