+++ tests/prelude/CMakeLists.txt
@@ -1,10 +1,10 @@
 cmake_minimum_required(VERSION 3.8)
-project(draft_tests CXX)
+project(prelude_tests CXX)
 
 # ${CMAKE_PROJECT_NAME}: project name of the root CMakeLists.txt
 # ${PROJECT_NAME}: project name of the current CMakeLists.txt
 add_executable(${PROJECT_NAME}
-  draft.cpp
+  prelude.cpp
 )
 
 target_link_libraries(${PROJECT_NAME} PRIVATE
