+++ tests/prelude/prelude.cpp
@@ -1,6 +1,6 @@
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
-#include <mpc/functional.hpp>
+#include <mpc/prelude.hpp>
 #include "../stdfundamental.hpp"
 
 using add_op = decltype([](const auto& x, const auto& y) { return x + y; });
@@ -11,7 +11,7 @@ struct add_t : mpc::perfect_forward<add_op> {
 
 inline constexpr add_t add;
 
-TEST_CASE("functional compose", "[functional][compose]") {
+TEST_CASE("prelude compose", "[prelude][compose]") {
   {
     const auto inc = [](auto x) { return x + 1; };
     const auto two = mpc::compose(inc, inc);
@@ -62,7 +62,7 @@ struct add3way_t : mpc::perfect_forward<add3way_op> {
 
 inline constexpr add3way_t add3way;
 
-TEST_CASE("functional flip", "[functional][flip]") {
+TEST_CASE("prelude flip", "[prelude][flip]") {
   {
     CHECK(mpc::flip % add % "A"s % "B"s == "BA"s);
     CHECK(mpc::flip % add3way % "A"s % "B"s % "C"s == "BAC"s);
